Implement pause, restart, deposit, withdraw APIs
Implement the hardhat tasks sab_pause, sab_restart, sab_deposit and
sab_withdraw. These invoke corresponding FlowStreamCreator methods,
so that all stream updates originate from that contract's address.
This way, for this prototype, the "sender" of all streams can be set
to the FlowStreamCreator.

Also implemented sab_mint and sab_balance, 2 APIs to operate the ERC20
balances of arbitrary addresses. sab_mint is useful to send coins to
the Creator contract address, before invoking depositFlowStream().
sab_balance can be used to check the contract address's balance after
a withdrawal to see the movement of funds.

Added constants in FlowStreamCreator with explanation of calculations
of rate per second. Added pauseFlowStream(), restartFlowStream(),
depositFlowStream() and withdrawFlowStream() methods.

Minor: added 2 convenience vars to setparsecvars.sh, not used in code
but can be retrieved from console if needed.

PS: in a production design, preferable to set sender of each stream
to be the actual originator (say an employer), so that deposits can
go directly from that sender to the Flow contract. But then pause/
restart would need to come directly from that address to the Flow
contract.